@@ -20,15 +20,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.12",
-    "@angular/common": "^17.3.12",
-    "@angular/compiler": "^17.3.12",
-    "@angular/core": "^17.3.12",
-    "@angular/forms": "^17.3.12",
-    "@angular/platform-browser": "^17.3.12",
-    "@angular/platform-browser-dynamic": "^17.3.12",
-    "@angular/router": "^17.3.12",
-    "@angular/service-worker": "^17.3.12",
+    "@angular/animations": "^18.2.13",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
+    "@angular/forms": "^18.2.13",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
+    "@angular/service-worker": "^18.2.13",
     "@ant-design/colors": "^7.1.0",
     "@auth0/angular-jwt": "^5.2.0",
     "@ngrx/effects": "^17.2.0",
@@ -48,15 +48,15 @@
     "zone.js": "^0.14.10"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.10",
+    "@angular-devkit/build-angular": "^18.2.14",
     "@angular-eslint/builder": "^19.0.2",
     "@angular-eslint/eslint-plugin": "^19.0.2",
     "@angular-eslint/eslint-plugin-template": "^19.0.2",
     "@angular-eslint/schematics": "^19.0.2",
     "@angular-eslint/template-parser": "^19.0.2",
-    "@angular/cli": "^17.3.10",
-    "@angular/compiler-cli": "^17.3.12",
-    "@angular/localize": "^17.3.12",
+    "@angular/cli": "^18.2.14",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/localize": "^18.2.13",
     "@schematics/angular": "^17.3.10",
     "@types/jasmine": "^4.6.4",
     "@types/luxon": "^3.4.2",
