@@ -9,7 +9,7 @@ import { fr_FR } from 'ng-zorro-antd/i18n';
 import { DatePipe, registerLocaleData } from '@angular/common';
 import fr from '@angular/common/locales/fr';
 import { ServiceWorkerModule } from '@angular/service-worker';
-import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HomeModule } from './modules/home/home.module';
 import { StoreModule } from '@ngrx/store';
@@ -32,34 +32,28 @@ const ngZorroConfig: NzConfig = {
   message: { nzMaxStack: 1, nzDuration: 2000 },
 };
 
-@NgModule({
-  declarations: [AppComponent, FourOhFourComponent],
-  imports: [
-    AppRoutingModule,
-    BrowserModule,
-    HttpClientModule,
-    BrowserAnimationsModule,
-    RouterModule,
-    SharedModule,
-    HomeModule,
-    LoginModule,
-    ServiceWorkerModule.register('ngsw-worker.js', {
-      enabled: environment.production,
-    }),
-    StoreModule.forRoot({}, {}),
-    EffectsModule.forRoot([]),
-    StoreDevtoolsModule.instrument({
-      maxAge: 25,
-      logOnly: environment.production,
-    }),
-  ],
-  providers: [
-    { provide: NZ_I18N, useValue: fr_FR },
-    { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
-    { provide: HTTP_INTERCEPTORS, useClass: HttpResponseInterceptor, multi: true },
-    { provide: NZ_CONFIG, useValue: ngZorroConfig },
-    DatePipe,
-  ],
-  bootstrap: [AppComponent],
-})
+@NgModule({ declarations: [AppComponent, FourOhFourComponent],
+    bootstrap: [AppComponent], imports: [AppRoutingModule,
+        BrowserModule,
+        BrowserAnimationsModule,
+        RouterModule,
+        SharedModule,
+        HomeModule,
+        LoginModule,
+        ServiceWorkerModule.register('ngsw-worker.js', {
+            enabled: environment.production,
+        }),
+        StoreModule.forRoot({}, {}),
+        EffectsModule.forRoot([]),
+        StoreDevtoolsModule.instrument({
+            maxAge: 25,
+            logOnly: environment.production,
+        })], providers: [
+        { provide: NZ_I18N, useValue: fr_FR },
+        { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
+        { provide: HTTP_INTERCEPTORS, useClass: HttpResponseInterceptor, multi: true },
+        { provide: NZ_CONFIG, useValue: ngZorroConfig },
+        DatePipe,
+        provideHttpClient(withInterceptorsFromDi()),
+    ] })
 export class AppModule { }
